add Db.js for dummy data for better overview in index.js, as well as minor changes. additionally an input field for a searchbar was hard coded.
