+++ components/SearchbarLostItem.js
@@ -1,0 +1,10 @@
+import React from "react";
+
+export default function SearchbarLostItem() {
+  return (
+    <div>
+      <h4>Search</h4>
+      <input placeholder="What are you looking for?"></input>
+    </div>
+  );
+}
