+++ pages/index.js
@@ -1,171 +1,11 @@
 import Head from "next/head";
 import styled from "styled-components";
-import LostItemPage from "../components/LostItemPage";
+import LostItem from "../components/LostItem";
+import SearchbarLostItem from "../components/SearchbarLostItem";
+import FoundItem from "../components/LostItem";
+import { lostItems, foundItems } from "../components/Db";
 
 export default function Home() {
-  const lostItems = [
-    {
-      name: "Key",
-      description: "Silver Key",
-      img: "/assets/lostItems/one.jpeg",
-      id: 1,
-    },
-    {
-      name: "Headphones",
-      description: "Silver and grey Headphones",
-      img: "/assets/lostItems/two.jpeg",
-      id: 2,
-    },
-    {
-      name: "Teapot",
-      description: "Red and orange Teapot",
-      img: "/assets/lostItems/three.jpeg",
-      id: 3,
-    },
-    {
-      name: "Lighter",
-      description: "White, silver and lightblue Lighter",
-      img: "/assets/lostItems/four.jpeg",
-      id: 4,
-    },
-    {
-      name: "Batteries",
-      description: "Yellow Batteries",
-      img: "/assets/lostItems/five.jpeg",
-      id: 5,
-    },
-    {
-      name: "Compass",
-      description: "Golden Compass",
-      img: "/assets/lostItems/six.jpeg",
-      id: 6,
-    },
-    {
-      name: "Bottle",
-      description: "Orange Bottle",
-      img: "/assets/lostItems/seven.jpeg",
-      id: 7,
-    },
-    {
-      name: "Camera",
-      description: "Black and silver analog Camera",
-      img: "/assets/lostItems/eight.jpeg",
-      id: 8,
-    },
-    {
-      name: "Backpack",
-      description: "Black and brown backpack found in the woods",
-      img: "/assets/lostItems/nine.jpeg",
-      id: 9,
-    },
-    {
-      name: "Camera lens",
-      description: "Black and silver camera lens",
-      img: "/assets/lostItems/ten.jpeg",
-      id: 10,
-    },
-    {
-      name: "Lollipop",
-      description: "Red lollipop",
-      img: "/assets/lostItems/eleven.jpeg",
-      id: 11,
-    },
-    {
-      name: "Band aid",
-      description: "Red and white band aid box",
-      img: "/assets/lostItems/twelve.jpeg",
-      id: 12,
-    },
-    {
-      name: "Kids bike",
-      description: "Blue, yellow, white and red kids bike",
-      img: "/assets/lostItems/thirteen.jpeg",
-      id: 13,
-    },
-  ];
-
-  const foundItems = [
-    {
-      name: "Camera",
-      description: "Camera underwater",
-      img: "/assets/foundItems/one.jpg",
-      id: 1,
-    },
-    {
-      name: "Backpack",
-      description: "Green and black backpack found in the woods",
-      img: "/assets/foundItems/two.jpg",
-      id: 2,
-    },
-    {
-      name: "Diamonds",
-      description: "We found a ton of diamonds!!!",
-      img: "assets/foundItems/three.jpeg",
-      id: 3,
-    },
-    {
-      name: "One shoe",
-      description: "Found one shoe. Where might the other one be?",
-      img: "/assets/foundItems/four.jpeg",
-      id: 4,
-    },
-    {
-      name: "Ring",
-      description: "Found the most beautiful ring with blue stone attached",
-      img: "/assets/foundItems/five.jpeg",
-      id: 5,
-    },
-    {
-      name: "Key",
-      description: "Ancient Key found in the Woods",
-      img: "/assets/foundItems/six.jpeg",
-      id: 6,
-    },
-    {
-      name: "Keyboard",
-      description: "Oldschool Keyboard with Keys missing",
-      img: "/assets/foundItems/seven.jpeg",
-      id: 7,
-    },
-    {
-      name: "Shoe",
-      description: "Found one shoe in an abandoned building",
-      img: "/assets/foundItems/eight.jpeg",
-      id: 8,
-    },
-    {
-      name: "Handshoe",
-      description: "One handshoe found on a parking bench",
-      img: "/assets/foundItems/nine.jpeg",
-      id: 9,
-    },
-    {
-      name: "Camera",
-      description: "Black camera found in the woods",
-      img: "/assets/foundItems/ten.jpeg",
-      id: 10,
-    },
-    {
-      name: "Football",
-      description: "Found a football",
-      img: "/assets/foundItems/eleven.jpeg",
-      id: 11,
-    },
-    {
-      name: "Golden coins",
-      description: "Found some golden coins. Wont give em back.",
-      img: "/assets/foundItems/twelve.jpeg",
-      id: 12,
-    },
-    {
-      name: "Golden Jewelry",
-      description:
-        "Found jewelry and intend to keep it. Just posting here for recognition.",
-      img: "/assets/foundItems/thirteen.jpeg",
-      id: 13,
-    },
-  ];
-
   return (
     <div>
       <Head>
@@ -176,13 +16,27 @@ export default function Home() {
 
       <Main>
         <h1>My App</h1>
+        <SearchbarLostItem />
         <ul>
           {lostItems.map((lostItem) => (
-            <LostItemPage
+            <LostItem
               name={lostItem.name}
               description={lostItem.description}
               img={lostItem.img}
+              key={lostItem.id}
               id={lostItem.id}
+            />
+          ))}
+        </ul>
+        <h2>Second Array of Dummy Data begins here:</h2>
+        <ul>
+          {foundItems.map((foundItem) => (
+            <FoundItem
+              name={foundItem.name}
+              description={foundItem.description}
+              img={foundItem.img}
+              key={foundItem.id}
+              id={foundItem.id}
             />
           ))}
         </ul>
